Create separate module for plugin based tasks
CLOUDBLD-6300

To prepare for a future where tasks do not have to be based on plugins,
create a special module for tasks that *are* based on them.

Move PluginsDef to this module. Make the the orchestrator, worker and
source build tasks inherit from the new PluginBasedTask.
